requirements.txt
discord==1.7.3
discord.py[voice]==1.7.3
python-dotenv==0.19.2
youtube-dl==2021.12.17